+++ Source/Program.cs
@@ -2,7 +2,9 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Aksio.IngressMiddleware;
+using Aksio.IngressMiddleware.BearerTokens;
 using Aksio.IngressMiddleware.Configuration;
+using Aksio.IngressMiddleware.Identities;
 using Aksio.IngressMiddleware.Tenancy;
 
 UnhandledExceptionsManager.Setup();
